Include kind of DateTime in messages (breaking change)
Previously, DateTime.Ticks was used to obtain an Int64 value for serialisation, and new DateTime(Int64) for deserialisation. However this method excludes the content of the DateTime.Kind field, which models UTC/local/unspecified.

DateTime internally is a 64-bit value, however it only uses 62 bits for ticks. The 2 MSb are flags for the kind.

With this change, we use ToBinary for serialisation and FromBinary for deserialisation. These also yield an Int64, but include the flags.

Compatibility between this and earlier versions of the library:

* Old serialisation with new deserialisation: Works as before. All received dates are considered as having unspecified time, which is the same behaviour as prior to this commit.

* New serialisation with old deserialisation: Fails only if the serialised date's kind is UTC or local. In such cases, the received value is treated as a tick count, which is out of range due to the flags, raising an ArgumentOutOfRangeException. However if the serialised date has unspecified kind, things continue to work as before.

This sort of breaking change will cease to occur once the library reaches version 1.0.